Partially upgrades to Molecule v2
This role has been gathering dust. We still use it in prod, but it
hasn't undergone any changes in nearly a year. The testing env was still
using Molecule v1, which has been deprecated for the better part of a
year now.

Upgraded the requirements to Molecule v2 and ported most of the config,
but wasn't able to pull in the proper dependencies. Not going to debug
that now: the role doesn't even have CI enabled.